Fix images deleting conditions
Null is strictly necessary when the entities are firstly created (do not have a ImageInfo instance initialized yet)

+++ backend/src/main/java/com/tfg/backend/controller/CategoryRestController.java
@@ -60,7 +60,7 @@ public class CategoryRestController {
         Category category = findCategoryHelper(id);
 
         // Delete the previous image from MinIO if it is not the default photo
-        if (!category.getCategoryImage().equals(GlobalDefaults.CATEGORY_IMAGE)){
+        if (category.getCategoryImage() != null && !category.getCategoryImage().equals(GlobalDefaults.CATEGORY_IMAGE)){
             storageService.deleteFile(category.getCategoryImage().getS3Key());
         }
